requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# RAG and AI
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.0
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0

# Web scraping and parsing
beautifulsoup4==4.12.2
requests==2.31.0
scrapy==2.11.0
selenium==4.15.0
playwright==1.40.0

# Data processing
pandas==2.1.3
numpy==1.25.0
scikit-learn==1.3.0

# Database
sqlalchemy==2.0.23
alembic==1.12.0
psycopg2-binary==2.9.7

# Utilities
pyyaml==6.0.1
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Security
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Git integration
GitPython==3.1.40